+++ players.js
@@ -1,0 +1,36 @@
+export const player1 = {
+    player: 1,
+    name: 'scorpion',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
+    weapon: ['knifw', 'sword', 'fire'],
+    changeHP,
+    elHP,
+    renderHP
+}
+
+export const player2 = {
+    player: 2,
+    name: 'subzero',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
+    weapon: ['knifw', 'sword', 'fire'],
+    changeHP,
+    elHP,
+    renderHP
+}
+
+function changeHP(hp) {
+    this.hp -= hp
+    if (this.hp < 0) {
+        this.hp = 0
+    }
+}
+
+function elHP() {
+    return document.querySelector(`.player${this.player} .life`)
+}
+
+function renderHP() {
+    this.elHP().style.width = `${this.hp}%`
+}
